Well as for the commit this time i have made some updates on the initialising the mutex part, i made a funtion to setup the writing forks and mutex forks, i tried adding like a dual while loop for the writing fork and the mutex fork but unfortunately, the mutex fork is the one that will be running again and again for every philo that has to be initialised but for the writing fork it only has to run once so i thought i dont need ot put a while loop there, plus i also used a typedef struct this time so i dont have to like call typedef struct s1 all the time but i can just use t_data. much more efecient.
finally before i finish i also added a bunch of test cases for the
initialising and multiple printfs to make sure tha i know where is each
part of the code being parsed every single step of the way. SO i caiterd
for 0 , 1,5,50,70,100,and a 1000 philo, so i tried handling for the edge
cases, and it works.(YAY!!) but besides that i guess everything is okay
and for the future dev who wish to read this i would like to say, make
sure you know how to use structs, what is pthread_mutex.