Commit for progress on 21/03/2025...init_philo!
for the init philo function, initialy i was pretty unfirmiliar with threads,
i mean i know how it works on the surface but not really good at using it well
thus what i did was i refered to some medium post, i would recomend this guy.
[https://medium.com/@ruinadd/philosophers-42-guide-the-dining-philosophers-problem-893a24bc0fe2#id_token=eyJhbGciOiJSUzI1NiIsImtpZCI6ImVlMTkzZDQ2NDdhYjRhMzU4NWFhOWIyYjNiNDg0YTg3YWE2OGJiNDIiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20iLCJhenAiOiIyMTYyOTYwMzU4MzQtazFrNnFlMDYwczJ0cDJhMmphbTRsamRjbXMwMHN0dGcuYXBwcy5nb29nbGV1c2VyY29udGVudC5jb20iLCJhdWQiOiIyMTYyOTYwMzU4MzQtazFrNnFlMDYwczJ0cDJhMmphbTRsamRjbXMwMHN0dGcuYXBwcy5nb29nbGV1c2VyY29udGVudC5jb20iLCJzdWIiOiIxMTExNDU0NDY5NTEzMzQ4NTc4NzQiLCJlbWFpbCI6InJhamF2aWtyYW1hbjIwMDZAZ21haWwuY29tIiwiZW1haWxfdmVyaWZpZWQiOnRydWUsIm5iZiI6MTc0MjY4NjI2MiwibmFtZSI6IlJBSkEgVklLUkFNQU4iLCJwaWN0dXJlIjoiaHR0cHM6Ly9saDMuZ29vZ2xldXNlcmNvbnRlbnQuY29tL2EvQUNnOG9jSU5DMi0yWE9ycEhRTndGMzF3ZkpBMm52bm5BUGU5VmFRZlRCY0tVZ2FkYnRvZU5sOD1zOTYtYyIsImdpdmVuX25hbWUiOiJSQUpBIiwiZmFtaWx5X25hbWUiOiJWSUtSQU1BTiIsImlhdCI6MTc0MjY4NjU2MiwiZXhwIjoxNzQyNjkwMTYyLCJqdGkiOiI2NmYwZjllZjlkOWFkNTA2Njc3MGE0MzQ3NjQ3MzdhYWUzMjZmYmVhIn0.llQOZv_N33evLsYr0xENFEchRf4v3uiZtxFmpPg6yn2Aw0K81djwzkHLUU7-EyxZroyvspHzTI3lxyl0uhZ87T2_eF5pURM_6BCrTUcidcLg1qsMjmpV4_mRlpDSTMf0zuIAmd4Wfg23ThtFUTNibeSWkXOSFLIGlDE8u2i7HRqWz5iABxDZ5pYrwlgVZIy9KxynNMXbQKF8pL_NC4PBs4pnBs3uwS3lLQVeKoIq0tTpR2fcJBjHLBe9l4NKVymH1cgTlEn0uWtB3zAmIQCr33O8BWKFrbFRnjEU1MMSwnU6uJ9c0IxRhFCoE-b8a7OVzXMLWOePcXlY_7MC4mmWsw]
this guys ruinadd, his explaination is pretty good. He gave a link, to understand
the basics of threads, and reenforced my current understanding.

Moreover, i also managed to make the programm to initialise all the variables to
the struct, for example-> the philos ID, left_fork ID, and so on.But the most chalanging
thing was to make sure that the sturcts work properly, moreover what i also came across
was the fact that the structs required the same variable declared in the previous
function which was the, sum_philo. As said by the great aristottle, When there is a will
there is a way. So what i did was i came across this unique program where they implimented
a struct in a struct, and when i did my study on them its said that what it does is that
it allows you to call one stuct and all the other variables from a diffrent sturct can still
accessed......